+++ client/admin-plugin-client-plugin.ts
@@ -208,8 +208,6 @@ function register ({ registerHook, registerSettingsScript, peertubeHelpers }: Re
           return options.formValues['chat-type'] !== ('external-uri' as ChatType)
         case 'chat-style':
           return options.formValues['chat-type'] === 'disabled'
-        case 'chat-only-locals-warning':
-          return options.formValues['chat-only-locals'] === true
         case 'chat-per-live-video-warning':
           return !(options.formValues['chat-all-lives'] === true && options.formValues['chat-per-live-video'] === true)
       }