+++ shared/lib/video.ts
@@ -1,7 +1,6 @@
 import { parseConfigUUIDs } from './config'
 
 interface SharedSettings {
-  'chat-only-locals': boolean
   'chat-per-live-video': boolean
   'chat-all-lives': boolean
   'chat-all-non-lives': boolean
@@ -27,12 +26,11 @@ interface SharedVideoBackend extends SharedVideoBase {
 type SharedVideo = SharedVideoBackend | SharedVideoFrontend
 
 function videoHasWebchat (settings: SharedSettings, video: SharedVideo): boolean {
-  if (settings['chat-only-locals']) {
-    if ('isLocal' in video) {
-      if (!video.isLocal) return false
-    } else {
-      if (video.remote) return false
-    }
+  // Never use webchat on remote videos.
+  if ('isLocal' in video) {
+    if (!video.isLocal) return false
+  } else {
+    if (video.remote) return false
   }
 
   if (settings['chat-per-live-video'] && video.isLive && video.pluginData && video.pluginData['livechat-active']) {